Splitting tests for travis
