+++ bigml/tests/test_35_compare_predictions.py
@@ -45,133 +45,6 @@ class TestComparePrediction(object):
         print("\nEnd of tests in: %s\n-------------------\n" % __name__)
 
 
-    def test_scenario1(self):
-        """
-            Scenario: Successfully comparing forecasts from time series:
-                Given I create a data source uploading a "<data>" file
-                And I wait until the source is ready less than <time_1> secs
-                And I create a dataset
-                And I wait until the dataset is ready less than <time_2> secs
-                And I create a time series with "<params>"
-                And I wait until the time series is ready less than <time_3> secs
-                And I create a local time series
-                When I create a forecast for "<input_data>"
-                Then the forecast is "<forecasts>"
-                And I create a local forecast for "<data_input>"
-                Then the local forecast is "<forecasts>"
-
-                Examples:
-                | data             | time_1  | time_2 | time_3 | input_data  | forecasts | params
-            ['data/grades.csv', '10', '10', '120', '{"000005": {"horizon": 5, "ets_models": {"names": ["A,Ad,N"], "criterion": "aic", "limit": 3}}}', '{"000005": [{"point_forecast":  [69.90959, 69.92755, 69.94514, 69.96236, 69.97922], "model": "A,Ad,N"}]}', '{"objective_fields": ["000001", "000005"]}'],
-
-
-        """
-        examples = [
-            ['data/grades.csv', '30', '30', '120', '{"000005": {"horizon": 5}}', '{"000005": [{"point_forecast": [73.96192, 74.04106, 74.12029, 74.1996, 74.27899], "model": "M,M,N"}]}', '{"objective_fields": ["000001", "000005"]}'],
-            ['data/grades.csv', '30', '30', '120', '{"000005": {"horizon": 5, "ets_models": {"names": ["M,N,N"], "criterion": "aic", "limit": 3}}}', '{"000005": [{"point_forecast":  [68.39832, 68.39832, 68.39832, 68.39832, 68.39832], "model": "M,N,N"}]}', '{"objective_fields": ["000001", "000005"]}'],
-            ['data/grades.csv', '30', '30', '120', '{"000005": {"horizon": 5, "ets_models": {"names": ["A,A,N"], "criterion": "aic", "limit": 3}}}', '{"000005": [{"point_forecast": [72.46247, 72.56247, 72.66247, 72.76247, 72.86247], "model": "A,A,N"}]}', '{"objective_fields": ["000001", "000005"]}'],
-            ['data/grades.csv', '30', '30', '120', '{"000005": {"horizon": 5}, "000001": {"horizon": 3, "ets_models": {"criterion": "aic", "limit": 2}}}', '{"000005": [{"point_forecast": [73.96192, 74.04106, 74.12029, 74.1996, 74.27899], "model": "M,M,N"}], "000001": [{"point_forecast": [55.51577, 89.69111, 82.04935], "model": "A,N,A"}, {"point_forecast": [56.67419, 91.89657, 84.70017], "model": "A,A,A"}]}', '{"objective_fields": ["000001", "000005"]}']]
-        show_doc(self.test_scenario1, examples)
-
-        for example in examples:
-            print("\nTesting with:\n", example)
-            source_create.i_upload_a_file(self, example[0])
-            source_create.the_source_is_finished(self, example[1])
-            dataset_create.i_create_a_dataset(self)
-            dataset_create.the_dataset_is_finished_in_less_than(self, example[2])
-            time_series_create.i_create_a_time_series_with_params(self, example[6])
-            time_series_create.the_time_series_is_finished_in_less_than(self, example[3])
-            time_series_create.create_local_time_series(self)
-            forecast_create.i_create_a_forecast(self, example[4])
-            forecast_create.the_forecast_is(self, example[5])
-            forecast_compare.i_create_a_local_forecast(self, example[4])
-            forecast_compare.the_local_forecast_is(self, example[5])
-
-
-    def test_scenario2(self):
-        """
-            Scenario: Successfully comparing forecasts from time series with "A" seasonality
-                Given I create a data source uploading a "<data>" file
-                And I wait until the source is ready less than <time_1> secs
-                And I create a dataset
-                And I wait until the dataset is ready less than <time_2> secs
-                And I create a time series with "<params>"
-                And I wait until the time series is ready less than <time_3> secs
-                And I create a local time series
-                When I create a forecast for "<input_data>"
-                Then the forecast is "<forecasts>"
-                And I create a local forecast for "<data_input>"
-                Then the local forecast is "<forecasts>"
-
-                Examples:
-                | data             | time_1  | time_2 | time_3 | input_data  | forecasts | params
-            ['data/grades.csv', '10', '10', '120', '{"000005": {"horizon": 5, "ets_models": {"names": ["A,Ad,A"], "criterion": "aic", "limit": 3}}}', '{"000005": [{"point_forecast":[66.16225, 72.17308, 66.65573, 73.09698, 70.51449], "model": "A,Ad,A"}]}', '{"objective_fields": ["000001", "000005"], "period": 12}']
-        """
-        examples = [
-
-            ['data/grades.csv', '30', '30', '120', '{"000005": {"horizon": 5}}', '{"000005": [{"point_forecast": [73.96192, 74.04106, 74.12029, 74.1996, 74.27899], "model": "M,M,N"}]}', '{"objective_fields": ["000001", "000005"], "period": 12}'],
-            ['data/grades.csv', '30', '30', '120', '{"000005": {"horizon": 5, "ets_models": {"names": ["M,N,A"], "criterion": "aic", "limit": 3}}}', '{"000005": [{"point_forecast":  [67.43222, 68.24468, 64.14437, 67.5662, 67.79028], "model": "M,N,A"}]}', '{"objective_fields": ["000001", "000005"], "period": 12}'],
-            ['data/grades.csv', '30', '30', '120', '{"000005": {"horizon": 5, "ets_models": {"names": ["A,A,A"], "criterion": "aic", "limit": 3}}}', '{"000005": [{"point_forecast": [74.73553, 71.6163, 71.90264, 76.4249, 75.06982], "model": "A,A,A"}]}', '{"objective_fields": ["000001", "000005"], "period": 12}']]
-        show_doc(self.test_scenario2, examples)
-
-        for example in examples:
-            print("\nTesting with:\n", example)
-            source_create.i_upload_a_file(self, example[0])
-            source_create.the_source_is_finished(self, example[1])
-            dataset_create.i_create_a_dataset(self)
-            dataset_create.the_dataset_is_finished_in_less_than(self, example[2])
-            time_series_create.i_create_a_time_series_with_params(self, example[6])
-            time_series_create.the_time_series_is_finished_in_less_than(self, example[3])
-            time_series_create.create_local_time_series(self)
-            forecast_create.i_create_a_forecast(self, example[4])
-            forecast_create.the_forecast_is(self, example[5])
-            forecast_compare.i_create_a_local_forecast(self, example[4])
-            forecast_compare.the_local_forecast_is(self, example[5])
-
-    def test_scenario3(self):
-        """
-            Scenario: Successfully comparing forecasts from time series with "M" seasonality
-                Given I create a data source uploading a "<data>" file
-                And I wait until the source is ready less than <time_1> secs
-                And I create a dataset
-                And I wait until the dataset is ready less than <time_2> secs
-                And I create a time series with "<params>"
-                And I wait until the time series is ready less than <time_3> secs
-                And I create a local time series
-                When I create a forecast for "<input_data>"
-                Then the forecast is "<forecasts>"
-                And I create a local forecast for "<data_input>"
-                Then the local forecast is "<forecasts>"
-
-                Examples:
-                | data             | time_1  | time_2 | time_3 | input_data  | forecasts | params
-
-,
-            ['data/grades.csv', '10', '10', '120', '{"000005": {"horizon": 5, "ets_models": {"names": ["M,Ad,M"], "criterion": "aic", "limit": 3}}}', '{"000005": [{"point_forecast": [73.75816, 74.60699, 66.71212, 72.49586, 71.76787], "model": "M,Ad,M"}]}', '{"objective_fields": ["000001", "000005"], "period": 12}'],
-            ['data/grades.csv', '10', '10', '120', '{"000005": {"horizon": 5, "ets_models": {"names": ["M,Md,M"], "criterion": "aic", "limit": 3}}}', '{"000005": [{"point_forecast": [74.3725, 75.02963, 67.15826, 73.19628, 71.66919], "model": "M,Md,M"}]}', '{"objective_fields": ["000001", "000005"], "period": 12}']
-
-        """
-        examples = [
-            ['data/grades.csv', '30', '30', '120', '{"000005": {"horizon": 5, "ets_models": {"names": ["M,N,M"], "criterion": "aic", "limit": 3}}}', '{"000005": [{"point_forecast":  [68.99775, 72.76777, 66.5556, 70.90818, 70.92998], "model": "M,N,M"}]}', '{"objective_fields": ["000001", "000005"], "period": 12}'],
-            ['data/grades.csv', '30', '30', '120', '{"000005": {"horizon": 5, "ets_models": {"names": ["M,A,M"], "criterion": "aic", "limit": 3}}}', '{"000005": [{"point_forecast": [70.65993, 78.20652, 69.64806, 75.43716, 78.13556], "model": "M,A,M"}]}', '{"objective_fields": ["000001", "000005"], "period": 12}'],
-            ['data/grades.csv', '30', '30', '120', '{"000005": {"horizon": 5, "ets_models": {"names": ["M,M,M"], "criterion": "aic", "limit": 3}}}', '{"000005": [{"point_forecast": [71.75055, 80.67195, 70.81368, 79.84999, 78.27634], "model": "M,M,M"}]}', '{"objective_fields": ["000001", "000005"], "period": 12}']]
-        show_doc(self.test_scenario3, examples)
-
-        for example in examples:
-            print("\nTesting with:\n", example)
-            source_create.i_upload_a_file(self, example[0])
-            source_create.the_source_is_finished(self, example[1])
-            dataset_create.i_create_a_dataset(self)
-            dataset_create.the_dataset_is_finished_in_less_than(self, example[2])
-            time_series_create.i_create_a_time_series_with_params(self, example[6])
-            time_series_create.the_time_series_is_finished_in_less_than(self, example[3])
-            time_series_create.create_local_time_series(self)
-            forecast_create.i_create_a_forecast(self, example[4])
-            forecast_create.the_forecast_is(self, example[5])
-            forecast_compare.i_create_a_local_forecast(self, example[4])
-            forecast_compare.the_local_forecast_is(self, example[5])
-
-
     def test_scenario4(self):
         """
             Scenario: Successfully comparing forecasts from time series with trivial models
